requirements.txt
flask
pymongo
python-dotenv
PyJWT
kafka-python
requests
prometheus-flask-exporter
prometheus_client
flask-cors
flask_limiter
flask_limiter.util
email-validator
slowapi